backend-llm/requirements.txt
fastapi
uvicorn
httpx
python-dotenv  
symspellpy
fuzzywuzzy
langdetect
python-Levenshtein